--- openstack/blockstorage/v1/apiversions/doc.go
@@ -1,3 +1,0 @@
-// Package apiversions provides information and interaction with the different
-// API versions for the OpenStack Block Storage service, code-named Cinder.
-package apiversions
