+++ openstack/blockstorage/v1/volumes/delete.go
@@ -1,0 +1,8 @@
+package volumes
+
+import "github.com/opentelekomcloud/gophertelekomcloud"
+
+func Delete(client *golangsdk.ServiceClient, id string) (err error) {
+	_, err = client.Delete(client.ServiceURL("volumes", id), nil)
+	return
+}
