--- openstack/blockstorage/v1/volumes/testing/doc.go
@@ -1,9 +1,0 @@
-// volumes_v1
-package testing
-
-/*
-This is package created is to hold fixtures (which imports testing),
-so that importing volumes package does not inadvertently import testing into production code
-More information here:
-https://github.com/rackspace/golangsdk/issues/473
-*/
