+++ openstack/blockstorage/v1/volumes/testing/requests_test.go
@@ -18,7 +18,7 @@ func TestList(t *testing.T) {
 
 	count := 0
 
-	_ = volumes.List(client.ServiceClient(), &volumes.ListOpts{}).EachPage(func(page pagination.Page) (bool, error) {
+	_ = volumes.List(client.ServiceClient(), volumes.ListOpts{}).EachPage(func(page pagination.Page) (bool, error) {
 		count++
 		actual, err := volumes.ExtractVolumes(page)
 		if err != nil {
@@ -53,7 +53,7 @@ func TestListAll(t *testing.T) {
 
 	MockListResponse(t)
 
-	allPages, err := volumes.List(client.ServiceClient(), &volumes.ListOpts{}).AllPages()
+	allPages, err := volumes.List(client.ServiceClient(), volumes.ListOpts{}).AllPages()
 	th.AssertNoErr(t, err)
 	actual, err := volumes.ExtractVolumes(allPages)
 	th.AssertNoErr(t, err)
@@ -79,7 +79,7 @@ func TestGet(t *testing.T) {
 
 	MockGetResponse(t)
 
-	actual, err := volumes.Get(client.ServiceClient(), "d32019d3-bc6e-4319-9c1d-6722fc136a22").Extract()
+	actual, err := volumes.Get(client.ServiceClient(), "d32019d3-bc6e-4319-9c1d-6722fc136a22")
 	th.AssertNoErr(t, err)
 
 	expected := &volumes.Volume{
@@ -118,11 +118,11 @@ func TestCreate(t *testing.T) {
 
 	MockCreateResponse(t)
 
-	options := &volumes.CreateOpts{
+	options := volumes.CreateOpts{
 		Size:             75,
 		AvailabilityZone: "us-east1",
 	}
-	n, err := volumes.Create(client.ServiceClient(), options).Extract()
+	n, err := volumes.Create(client.ServiceClient(), options)
 	th.AssertNoErr(t, err)
 
 	th.AssertEquals(t, n.Size, 4)
@@ -136,7 +136,7 @@ func TestDelete(t *testing.T) {
 	MockDeleteResponse(t)
 
 	res := volumes.Delete(client.ServiceClient(), "d32019d3-bc6e-4319-9c1d-6722fc136a22")
-	th.AssertNoErr(t, res.Err)
+	th.AssertNoErr(t, res)
 }
 
 func TestUpdate(t *testing.T) {
@@ -146,7 +146,7 @@ func TestUpdate(t *testing.T) {
 	MockUpdateResponse(t)
 
 	options := volumes.UpdateOpts{Name: "vol-002"}
-	v, err := volumes.Update(client.ServiceClient(), "d32019d3-bc6e-4319-9c1d-6722fc136a22", options).Extract()
+	v, err := volumes.Update(client.ServiceClient(), "d32019d3-bc6e-4319-9c1d-6722fc136a22", options)
 	th.AssertNoErr(t, err)
 	th.CheckEquals(t, "vol-002", v.Name)
 }
