--- openstack/blockstorage/v1/volumetypes/testing/doc.go
@@ -1,2 +1,0 @@
-// volumetypes_v1
-package testing
